==> 4_composition.Rmd <==
---
title: "Composition"
---

```{r}
#libraries
library(phyloseq)
library(vegan)
library(dplyr)
library(ggplot2)
library(reshape2)
library(genefilter)
#library(ggh4x) # devtools::install_github("teunbrand/ggh4x")
library(data.table)
```

```{r}
# seed and pathways
set.seed(0201)
sessionInfo()
filepath = "~/Library/CloudStorage/OneDrive-NTNU/Doktorgrad/Ferdige prosjekter/Project 3 MMS_FTS_r_K_experiment/MMS-FTS Overview results and manuscript plan/R_Analysis_MMS-FTS"
filepath_results =paste0( filepath, "/results/", Sys.Date(), "_")

filepath_figures = paste0(filepath, "/figures/" , Sys.Date(), "_")
full_experiment = readRDS(file = paste0(filepath,"/data/zOTUs/2022-01-11_zOTU_raw_dataset.rds") ) 
source(paste0(filepath, "/revised_figures_scripts/plot.settings.R"))
source(paste0(filepath, "/functions/min.mean.sd.max.R"))
source(paste0(filepath, "/functions/summarize_taxa_functions.R"))
```

```{r}
#attach data set
exp_nor = readRDS(file = paste0(filepath,"/results/2022-01-11_zOTU_dataset_scaled_rarified.rds") )
```

# Class
```{r}
# for plotting purposes we want to work with compositional data (scale in percentage)
my.physeq = microbiome::transform(x = exp_nor, transform = "compositional")

tax_rank = "Class"
min_percentage = 5
minTotRelAbun = min_percentage/100
nsamples_above = 2 

df_keep_taxa = data.frame(summarize_taxa(physeq = my.physeq, Rank = tax_rank, GroupBy = "Sample"))
df_keep_taxa = df_keep_taxa %>% group_by(df_keep_taxa[1]) %>% 
  summarise(in_n_samples = n(), 
            n_samples_above_min_ab = sum(totalA > minTotRelAbun ) ) %>%
  filter(n_samples_above_min_ab >= nsamples_above )

taxa_summary = data.frame(summarize_taxa(my.physeq, tax_rank, GroupBy = c("Sample")))
taxa_summary_df = merge(taxa_summary, data.frame(sample_data(my.physeq)), by = "Sample")
taxa_summary_df_filtered = taxa_summary_df %>% filter(Class %in% df_keep_taxa$Class ) 

taxa_summary_df_filtered_out = taxa_summary_df %>% filter(!Class %in% df_keep_taxa$Class ) %>%
  group_by(Sample, Sample_seq, DPH, Tank, Start_system, Current_system, Overall_system, cc, cc_s, Current_water) %>%
  summarise(totalA = sum(totalA)) %>%
  mutate(Class = paste0("a Less than ", min_percentage, "%")) 

taxa_summary_df_filtered_both = bind_rows(taxa_summary_df_filtered, taxa_summary_df_filtered_out)
```

```{r}
taxa_summary_df_filtered_both %>% mutate(totalA = totalA*100)   %>%
  group_by(Class) %>% 
  summarise(mean_ab = round( mean(totalA), 1) , 
            sd_ab = round( sd(totalA) , 1) )  
```

# Order
stronger filtering - 20% 2 samples
```{r}
# for plotting purposes we want to work with compositional data (scale in percentage)
my.physeq = microbiome::transform(x = exp_nor, transform = "compositional")

tax_rank = "Order"
min_percentage = 20
minTotRelAbun = min_percentage/100
nsamples_above = 2 

df_keep_taxa = data.frame(summarize_taxa(physeq = my.physeq, Rank = tax_rank, GroupBy = "Sample"))
df_keep_taxa = df_keep_taxa %>% group_by(df_keep_taxa[1]) %>% 
  summarise(in_n_samples = n(), 
            n_samples_above_min_ab = sum(totalA > minTotRelAbun ) ) %>%
  filter(n_samples_above_min_ab >= nsamples_above )

taxa_summary = data.frame(summarize_taxa(my.physeq, tax_rank, GroupBy = c("Sample")))
taxa_summary_df = merge(taxa_summary, data.frame(sample_data(my.physeq)), by = "Sample")
taxa_summary_df_filtered = taxa_summary_df %>% filter(Order %in% df_keep_taxa$Order ) 

taxa_summary_df_filtered_out = taxa_summary_df %>% filter(!Order %in% df_keep_taxa$Order ) %>%
  group_by(Sample, Sample_seq, DPH, Tank, Start_system, Current_system, Overall_system, cc, cc_s, Current_water) %>%
  summarise(totalA = sum(totalA)) %>%
  mutate(Order = paste0("a Less than ", min_percentage, "%")) 

taxa_summary_df_filtered_both = bind_rows(taxa_summary_df_filtered, taxa_summary_df_filtered_out)
```

```{r}
dfplot = taxa_summary_df_filtered_both %>% 
  mutate(totalA = totalA*100,
         DPH = as.character(DPH))   %>%
  group_by(Start_system, Order, Overall_system, DPH ) %>% 
  summarise(mean_ab = round( mean(totalA), 1) , 
            sd_ab = round( sd(totalA) , 1) ) %>%
  filter(Order != paste0("a Less than ", min_percentage, "%")) 

#facetted by initial treatment
dfplot %>% 
  ggplot2::ggplot(aes(x=DPH, weight = mean_ab, ymin=mean_ab-sd_ab, ymax=mean_ab+sd_ab, 
                      fill = Overall_system, group = Overall_system)) + 
     geom_bar      (position=position_dodge2(preserve = "single"), aes(y=mean_ab), stat="identity", color = "black") +
     geom_errorbar (position=position_dodge(width=0.9), colour="black", width = 0.2) +
  my.theme + 
  ggh4x::facet_nested(Order ~ Start_system, scales = "free", space = "free_x") +
  theme(panel.spacing = unit(0,"mm"), 
        strip.text.y = element_text(angle = 0), 
        text = element_text(size = 10 )) + 
  scale_fill_manual(values = system_overall_fill) + 
  geom_text(aes(y = mean_ab, label = paste0( round(mean_ab,0), "%") ) , vjust = -1, size = 3,  
                position = position_dodge2(preserve = "single", width = 0.9) )  + 
  ylab("Average relative abundance") 
```

```{r}
#facetted by samplind day
dfplot %>% 
  ggplot2::ggplot(aes(x=Overall_system, weight = mean_ab, ymin=mean_ab-sd_ab, ymax=mean_ab+sd_ab, 
                      fill = Overall_system, group = Overall_system)) + 
     geom_bar      (position=position_dodge2(preserve = "single"), aes(y=mean_ab), stat="identity", color = "black", width = 0.8) +
     geom_errorbar (position=position_dodge(width=0.9), colour="black", width = 0.2) +
  
  my.theme + 
  
  ggh4x::facet_nested(Order ~ DPH, scales = "free_x", space = "free_x") +
  
  theme(panel.spacing = unit(0,"mm"), 
        strip.text.y = element_text(angle = 0), 
        text = element_text(size = 10 )) + 
  scale_fill_manual("Rearing treatment", values = system_overall_fill) + 
  geom_text(aes(y = mean_ab, label = paste0( round(mean_ab,0), "%") ) , vjust = -1, size = 3,  
                position = position_dodge2(preserve = "single", width = 0.9) )  + 
  ylab("Average relative abundance") 

ggsave(filename = paste0(filepath_figures, "key_orders1.svg"), width = 180, height = 100, units = "mm")
```
