==> 6_flow_cytometry.Rmd <==
---
title: "Flow cytometry"
---

# Packages and filepath
```{r}
#libraries
library(dplyr)
library(ggplot2)
library(readxl)
#library(GGally) # install.packages("GGally")
library(svglite)
```

```{r}
# seed and pathways
set.seed(0201)
sessionInfo()
filepath = "~/Library/CloudStorage/OneDrive-NTNU/Doktorgrad/Ferdige prosjekter/Project 3 MMS_FTS_r_K_experiment/MMS-FTS Overview results and manuscript plan/R_Analysis_MMS-FTS/"
filepath_results = paste0(filepath, "results/", Sys.Date(), "_")
filepath_figures = paste0(filepath, "figures/",  Sys.Date(), "_")
source(paste0(filepath, "/revised_figures_scripts/plot.settings.R"))
source(paste0(filepath, "/functions/min.mean.sd.max.R"))
```

# Flow data
```{r}
flow = readxl::read_excel(paste0(filepath, "data/2021-06-28_All_results.xlsx"), 
                             sheet = "flowcytometry")
flow.1 = flow %>% dplyr::select(water_type, tank_number, incubation, millionbacteriaperml, DPH, water_fed)
tank_groups = readxl::read_excel(paste0(filepath, "data/2021-06-28_All_results.xlsx"), 
                             sheet = "tank_selection")
tank_groups = tank_groups %>% mutate(System_start = substr(Overall_system, start = 1, stop = 4) )
flow.2 = merge(flow.1, tank_groups, by = "tank_number", all = TRUE)
flow.in = flow.2 %>% dplyr::filter(water_type == "incoming")
flow.ta = flow.2 %>% dplyr::filter(water_type != "incoming") %>% 
  dplyr::mutate(Overall_system = case_when(DPH < 10 ~ System_start,  T ~ Overall_system))
flow.in = flow.in %>% mutate(System_start = tank_number)
flow.in.FTS = flow.in %>% dplyr::filter(System_start == "FTS")
flow.in.MMS = flow.in %>% dplyr::filter(System_start != "FTS")
flow.in.FTS1 = flow.in.FTS %>% dplyr::mutate(System_start = paste0(System_start, "+"))
flow.in.FTS2 = flow.in.FTS %>% dplyr::mutate(System_start = paste0(System_start, "-"))
flow.in.1 = dplyr::bind_rows(flow.in.MMS, flow.in.FTS1) 
flow.in.1 = dplyr::bind_rows(flow.in.1, flow.in.FTS2) 
flow.in.1 = flow.in.1 %>% dplyr::mutate(System_start = case_when(System_start == "MMS" ~ "MMS-", 
                                                                 T ~ System_start)) %>%
  mutate(Overall_system = paste0("x_InCm ", System_start)) 
flow.3 = dplyr::bind_rows(flow.ta, flow.in.1) 
flow.3 = flow.3 %>% dplyr::mutate(Sys.start = paste0("Start: ", System_start))

flow.3 = flow.3 %>% mutate(Overall_system = case_when( substring(Overall_system, 1, 1) == "x" ~ "Incoming water", 
                                              TRUE ~ Overall_system))
```

# Bacterial density
```{r}
flow.density = flow.3 %>% 
  dplyr::mutate(DPH = factor(DPH, levels = c("2", "9", "11", "15") ) ) %>%
  dplyr::filter(incubation == "before")
```
## Density plot
```{r}
flow.density %>%
  ggplot2::ggplot(aes(x=DPH, y = millionbacteriaperml )) + 
  my.theme + 
  facet_grid(~System_start) + 
  #rearing water
  geom_point(size = 3, 
             data =  flow.density %>% filter(water_type != "incoming"), 
             aes(x=DPH, y = millionbacteriaperml, 
                 group = interaction(Overall_system,DPH), 
                 shape = Overall_system, fill = Overall_system), 
             color = "black") + 
  
  #incoming
  geom_point(size = 3, 
             data =  flow.density %>% filter(water_type == "incoming"), 
             aes(x=DPH, y = millionbacteriaperml, 
                 group = interaction(Overall_system,DPH), 
                 shape = Overall_system, fill = Overall_system), 
             color = "black") + 
  
  scale_fill_manual("Rearing treatment", values = system_overall_fill_flow) + 
  scale_shape_manual("Rearing treatment", values = system_overall_shape_flow  )  + 
  
  scale_y_continuous(trans = "log10") +
  ylab(expression(paste("Bacterial density [",10^6, " cells ", mL^-1, "]"))) + 
  labs(caption = "Initial treatment") 

ggsave(filename = paste0(filepath_figures, "bacterial_density_days1.svg"), 
       width = 180, height = 90, units = "mm")
```

## density comparisons
```{r}
#high vs low cc overall 
cc_av = flow.density %>%
  mutate(cc = substr(System_start, 4,4)) %>%
  filter (water_type == "tank") %>% 
  group_by(water_type, cc) %>%
  summarise(av = mean(millionbacteriaperml)) 
cc_av$av[2]/cc_av$av[1] #7.8x

flow.density %>%
  mutate(cc = substr(System_start, 4,4)) %>%
  filter (water_type == "tank") %>% 
  group_by(water_type, cc, DPH) %>%
  rstatix::shapiro_test(millionbacteriaperml) 
#density was not normally distributed for each rearing system (p < 0.05), assessed by Shapiro-Wilk’s test.
```

```{r}
#high vs low cc day 2
cc_av_d2 = flow.density %>%
  mutate(cc = substr(System_start, 4,4)) %>%
  filter (water_type == "tank", DPH == 2) %>% 
  group_by(water_type, cc) %>%
  summarise(av = mean(millionbacteriaperml)) 
cc_av_d2$av[2]/cc_av_d2$av[1] #34.8x

#flinger test the null hypothesis that the variance is similar within each group. 
fligner.test(data = flow.density %>%
               mutate(cc = substr(System_start, 4,4)) %>%
               filter (water_type == "tank", DPH == 2) , 
             millionbacteriaperml ~ Overall_system) #0.005327 - variance not homogeneous

#FTS+
kruskal.test(formula = millionbacteriaperml ~ cc, 
             data = flow.density %>%
               mutate(cc = substr(System_start, 4,4)) %>%
               filter (water_type == "tank", DPH == 2)  )

#Kruskal-Wallis chi-squared = 11.327, df = 1, p-value = 0.0007637
```

```{r}
#high vs low cc day 9
cc_av_d9 = flow.density %>%
  mutate(cc = substr(System_start, 4,4)) %>%
  filter (water_type == "tank", DPH == 9) %>% 
  group_by(water_type, cc) %>%
  summarise(av = mean(millionbacteriaperml)) 
cc_av_d9$av[2]/cc_av_d9$av[1] #9.0689

#flinger test the null hypothesis that the variance is similar within each group. 
fligner.test(data = flow.density %>%
               mutate(cc = substr(System_start, 4,4)) %>%
               filter (water_type == "tank", DPH == 9) , 
             millionbacteriaperml ~ Overall_system) #0.009211 - variance not homogeneous

#FTS+
kruskal.test(formula = millionbacteriaperml ~ cc, 
             data = flow.density %>%
               mutate(cc = substr(System_start, 4,4)) %>%
               filter (water_type == "tank", DPH == 9)  )

#Kruskal-Wallis chi-squared = 11.446, df = 1, p-value = 0.0007166
```

## switching effect density 
```{r}
#averages per day
cc_av_switch_tank = flow.density %>%
  mutate(cc = substr(System_start, 4,4)) %>%
  filter ( DPH %in% c(11,15)) %>% 
  group_by(DPH, Sys.start, Overall_system, water_type) %>%
  summarise(av = mean(millionbacteriaperml)) 

#averages period after switch
cc_av_switch_tank_av = flow.density %>%
  mutate(cc = substr(System_start, 4,4)) %>%
  filter ( DPH %in% c(11,15)) %>% 
  group_by(Sys.start, Overall_system, water_type) %>%
  summarise(av = mean(millionbacteriaperml)) 

## ratios between switched and unswitched
#FTS+
  #11 DPH
  3.30/1.85 #1.783784
  #average 11 + 15 DPH
  3.350/2.425 #1.381443
#FTS-
  #11 DPH
  3.30/0.95 #3.47
  #average 11 + 15 DPH
  3.750/1.375 #2.72
#MMS+
  #11 DPH
  24.55/4.40 #5.579545
  #average 11 + 15 DPH
  18.975/4.850 #3.912371
#MMS-
  #11 DPH
  1.05/0.95 #1.105263
  #average 11 + 15 DPH
  2.100/1.225 #1.714286

#ratio water and incoming water
# FTS- 1.751592
  1.375/0.785
# FTS- to MMS- 8.241758
  3.750/0.455
# FTS+  3.089172
  2.425/0.785
# FTS+ to MMS+ 1.419492
  3.350/2.360
# MMS- 4.615385
  2.100/0.455
# MMS- to FTS- 1.56051
  1.225/0.785
# MMS+ 2.055085
  4.850/2.360
# MMS+ to FTS+  24.17197
  18.975/0.785
```

```{r}
den_af_sw = flow.density %>%
filter ( DPH %in% c(11,15), water_type == "tank") 
den_af_sw %>% 
  group_by(System_start) %>%
  rstatix::shapiro_test(millionbacteriaperml) 
#density was not normally distributed for each rearing system (p < 0.05), assessed by Shapiro-Wilk’s test.

#flinger test the null hypothesis that the variance is similar within each group. 
fligner.test(data = den_af_sw %>% filter(System_start == "FTS+"), 
             millionbacteriaperml ~ Overall_system) #0.02535 - variance not homogeneous
fligner.test(data = den_af_sw %>% filter(System_start == "FTS-"), 
             millionbacteriaperml ~ Overall_system) #0.7447 - variance homogeneous
fligner.test(data = den_af_sw %>% filter(System_start == "MMS+"), 
             millionbacteriaperml ~ Overall_system) #0.02491 - variance not homogeneous
fligner.test(data = den_af_sw %>% filter(System_start == "MMS-"), 
             millionbacteriaperml ~ Overall_system) #0.1842 - variance homogeneous
#many groups do not have normally distributed values, and are not homogenous in variance
#FTS+
kruskal.test(formula = millionbacteriaperml ~ Overall_system, data = den_af_sw %>% filter(System_start == "FTS+") ) 
#Kruskal-Wallis chi-squared = 0.18976, df = 1, p-value = 0.6631
#FTS-
kruskal.test(formula = millionbacteriaperml ~ Overall_system, data = den_af_sw %>% filter(System_start == "FTS-") ) 
#Kruskal-Wallis chi-squared = 5.3333, df = 1, p-value = 0.02092
#MMS+
kruskal.test(formula = millionbacteriaperml ~ Overall_system, data = den_af_sw %>% filter(System_start == "MMS+") ) 
#Kruskal-Wallis chi-squared = 5.3976, df = 1, p-value = 0.02016
#MMS-
kruskal.test(formula = millionbacteriaperml ~ Overall_system, data = den_af_sw %>% filter(System_start == "MMS-") ) 
#Kruskal-Wallis chi-squared = 0.0875, df = 1, p-value = 0.7674
```

#bacterial growth pot
```{r}
flow.pot = flow.3 %>% 
  dplyr::mutate(DPH = factor(DPH, levels = c("2", "9", "11", "15") ) ) %>%
  dplyr::filter(water_fed == "no") %>% 
  dplyr::group_by(DPH, tank_number, Overall_system, System_start, Sys.start, water_type)  

flow.pot.1 = reshape2::dcast(flow.pot, 
                             DPH+tank_number+water_type+Overall_system+System_start+Sys.start ~incubation, 
                value.var= "millionbacteriaperml", fun = sum, margins="incubation") %>%
  dplyr::mutate(gp = log10(after/before)) %>%
  dplyr::mutate(at_cc = case_when( gp>0 ~ 0,
                                 gp <= 0 ~ 1))
flow.pot.1_cc = flow.pot.1 %>% group_by(DPH, Overall_system, System_start, Sys.start, water_type) %>%
  summarise(cc_count = sum(at_cc), ntanks=n()) %>% 
  mutate(percent_cc = cc_count*100/ntanks) 
```


```{r}
flow.pot.1 %>%
   ggplot2::ggplot(aes(x=DPH, y = gp)) + 
  my.theme + 
  facet_grid(~System_start) + 
  #rearing water
  geom_point(size = 3, 
             data =  flow.pot.1 %>% filter(water_type != "incoming"), 
             aes(x=DPH, y = gp, 
                 group = interaction(Overall_system,DPH), 
                 shape = Overall_system, fill = Overall_system), 
             color = "black") + 
  
  #incoming
  geom_point(size = 3, 
             data =  flow.pot.1 %>% filter(water_type == "incoming"), 
             aes(x=DPH, y = gp, 
                 group = interaction(Overall_system,DPH), 
                 shape = Overall_system, fill = Overall_system), 
             color = "black") + 
  
  scale_fill_manual("Rearing treatment", values = system_overall_fill_flow) + 
  scale_shape_manual("Rearing treatment", values = system_overall_shape_flow  )  + 

  ylab("Bacterial net growth potential") + 
  labs(caption = "Initial treatment")

ggsave(filename = paste0(filepath_figures, "bacterial_growthpot.svg"), 
       width = 180, height = 90, units = "mm")
```

```{r}
gp_with_current = flow.pot.1 %>%
  filter(water_type == "tank") %>%
  mutate(current_system = case_when(DPH %in% c(2, 9) ~ System_start, 
                                    Overall_system == System_start ~ System_start, 
                                    T ~ substr(Overall_system, 9,12)))
gp_with_current %>% 
  group_by(water_type, current_system) %>%
  summarise(aver_gp = mean(gp))
```
