==> 7_survival.Rmd <==
---
title: "Survival main tanks"
---

```{r}
#libraries
library(dplyr)
library(ggplot2)
library(readxl)
```

```{r}
# seed and pathways
set.seed(0201)
sessionInfo()
filepath = "~/Library/CloudStorage/OneDrive-NTNU/Doktorgrad/Ferdige prosjekter/Project 3 MMS_FTS_r_K_experiment/MMS-FTS Overview results and manuscript plan/R_Analysis_MMS-FTS"
filepath_results =paste0( filepath, "/results/", Sys.Date(), "_")
filepath_figures = paste0(filepath, "/figures/" , Sys.Date(), "_")

source(paste0(filepath, "/revised_figures_scripts/plot.settings.R"))
source(paste0(filepath, "/revised_figures_scripts/min.mean.sd.max.R"))
```

```{r}
#attach data set
survival = readxl::read_excel(paste0(filepath, "/data/2021-06-28_All_results.xlsx"), 
                             sheet = "survival_maintank")
```

```{r}
tank_groups = readxl::read_excel(paste0(filepath, "/data/2021-06-28_All_results.xlsx"), 
                             sheet = "tank_selection")
tank_groups = tank_groups %>% mutate(System_start = substr(Overall_system, start = 1, stop = 4) ) 
```

```{r}
survival.1 = merge(survival, tank_groups)
survival.2 = survival.1 %>% mutate(survival = round(Survival_n_larvea*100/(100*100),1) )  
avg.survival = survival.2 %>% group_by(Overall_system) %>% summarise(avg_sur = mean(survival)) %>% 
  mutate(sur_t = paste0(round(avg_sur,0), "%"))
survival.3 = merge(survival.2, avg.survival)
```

Survival plot
```{r}
survival.3 %>% 
  ggplot2::ggplot(aes(x = Overall_system, y =survival )) + 
  facet_grid(~System_start, scales = "free_x") + 
  
  my.theme + 
  scale_fill_manual("Rearing treatment",values = system_overall_fill) + 
  scale_color_manual("Rearing treatment",values = system_overall_fill) + 
  scale_shape_manual("Rearing treatment",values = system_overall_shape_21) +  
  
  stat_summary(fun.y=mean, geom="bar", fill = "grey90", color = "black") +

  geom_point(size = 4, alpha  =0.7,
             aes(shape = Overall_system, fill = Overall_system), color = "black") + 
  
  geom_text(aes(label = round(survival)), size = 2) + 
    stat_summary(fun.y=mean, geom="text", aes(label=paste0(sprintf("%1.0f", ..y..),"%")), 
               size=4, show.legend=FALSE,   hjust = 1.1, vjust= -0.5 ) +
  theme(legend.position = "none") + 

  coord_cartesian(expand = TRUE) + 
  xlab("") + 
  ylab("Survival (%)") + 
  theme(text = element_text(size = 12)) + 
  theme(axis.text.x = element_text(size = 9)) 

```
```{r}
ggsave(filename = paste0(filepath_figures, "final_survival_bar_s.svg"), width = 180, height = 80, units = "mm")
```

```{r}
saveRDS(object = survival.3, file = paste0(filepath_results, "survival.RDS"))
```
